Split up parser code
