+++ pkg/parser/parser_test.go
@@ -1,11 +1,9 @@
 package parser
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/dharmab/skyeye/pkg/brevity"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,36 +12,6 @@ const TestCallsign = "Skyeye"
 type parserTestCase struct {
 	text     string
 	expected any
-}
-
-func TestParsePilotCallsign(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		name     string
-		expected string
-	}{
-		{"jeff", "jeff"},
-		{"Jeff", "jeff"},
-		{"Wardog 1 4", "wardog 1 4"},
-		{"Wardog 14", "wardog 1 4"},
-		{"Wardog 1-4", "wardog 1 4"},
-		{"WARDOG 14", "wardog 1 4"},
-		{"WARDOG14", "wardog 1 4"},
-		{"Mobius 1", "mobius 1"},
-		{"Red 243", "red 2 4 3"},
-		{"Red 054", "red 0 5 4"},
-		{"Gunfighter request", "gunfighter"},
-		{"This is Red 7", "red 7"},
-	}
-
-	for _, test := range testCases {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			actual, ok := ParsePilotCallsign(test.name)
-			require.True(t, ok)
-			assert.Equal(t, test.expected, actual)
-		})
-	}
 }
 
 func runParserTestCases(
@@ -203,23 +171,4 @@ func TestParserCheckIn(t *testing.T) {
 		actual := request.(*brevity.CheckInRequest)
 		require.Equal(t, expected.Callsign, actual.Callsign)
 	})
-}
-
-func TestIsSimilar(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		a        string
-		b        string
-		expected bool
-	}{
-		{"SkyEye", "Sky Eye", true},
-		{"Bandar", "Bandog", true},
-		{"Sky Eye", "Ghost Eye", false},
-	}
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("%s_%s", test.a, test.b), func(t *testing.T) {
-			t.Parallel()
-			assert.Equal(t, test.expected, isSimilar(test.a, test.b))
-		})
-	}
 }
